move files to components folder
