+++ apps/consent-ui/src/components/views/Dashboard/Dashboard.module.scss
