+++ apps/consent-ui/src/components/views/Dashboard/index.tsx
@@ -24,7 +24,7 @@ import Card from 'src/components/common/Card';
 import { getTranslation, ValidLanguage } from 'src/i18n';
 import ConsentsImage from 'src/public/consents.jpeg';
 
-import LocalizedLink from '../Link/LocalizedLink';
+import LocalizedLink from '../../common/Link/LocalizedLink';
 
 import styles from './Dashboard.module.scss';
